No grid, no heat map
Even using your code I’m not getting a heat map or the gridlines. The
one thing I’ve changed is the password so that it can enter the
database. I don’t think I’ll be able to do any of the calculations if
nothing shows up in the first place. Can we please go over this in
class?